LUCENE-1257: Cleanup unneeded casts
git-svn-id: https://svn.apache.org/repos/asf/lucene/java/trunk@827811 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/lucene/search/DefaultSimilarity.java
@@ -37,7 +37,7 @@ public class DefaultSimilarity extends Similarity {
       numTerms = state.getLength() - state.getNumOverlap();
     else
       numTerms = state.getLength();
-    return (float) (state.getBoost() * lengthNorm(field, numTerms));
+    return (state.getBoost() * lengthNorm(field, numTerms));
   }
   
   /** Implemented as <code>1/sqrt(numTerms)</code>. */
